+++ src/services/chain/nft/sweet-and-sour/types.ts
@@ -1,0 +1,4 @@
+export type SweetAndSourData = {
+  totalAmount: string;
+  totalClaimed: string;
+};
