+++ src/services/chain/nft/unexpected-move/types.ts
@@ -1,0 +1,5 @@
+export type UnexpectedMoveData = {
+  totalAmount: string;
+  totalClaimed: string;
+  totalExchanged: string;
+};
