+++ src/store/modules/nft/getters.ts
@@ -1,6 +1,6 @@
 import { GetterTree } from 'vuex';
-import { RootStoreState } from '@/store/types';
 
+import { RootStoreState } from '@/store/types';
 import { NFTStoreState } from './types';
 
 export default {} as GetterTree<NFTStoreState, RootStoreState>;
